+++ src/utils/flux/store/__spec__.js
@@ -126,6 +126,20 @@ describe('Store', () => {
           });
         });
       });
+
+      describe('when an invalid action is dispatched', () => {
+        it('throws an error for empty string', () => {
+          expect(function() { instance.dispatcherCallback({ actionType: '' }) }).toThrowError("You are subscribed to an invalid action (maybe the constant is incorrect or missing). Check Store.");
+        });
+
+        it('throws an error for null', () => {
+          expect(function() { instance.dispatcherCallback({ actionType: null }) }).toThrowError("You are subscribed to an invalid action (maybe the constant is incorrect or missing). Check Store.");
+        });
+
+        it('throws an error for undefined', () => {
+          expect(function() { instance.dispatcherCallback({ actionType: undefined }) }).toThrowError("You are subscribed to an invalid action (maybe the constant is incorrect or missing). Check Store.");
+        });
+      });
     });
 
     describe('undo', () => {